tool_calling/pyproject.toml
[project]
name = "tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
